backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.0
pytest==8.2.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4